requirements.txt
panda3d
panda3d-simplepbr
wecs
